Moved ParameterSet headers to src if not needed for API
