+++ FWCore/ParameterSet/src/ANDGroupDescription.cc
@@ -1,5 +1,5 @@
 
-#include "FWCore/ParameterSet/interface/ANDGroupDescription.h"
+#include "FWCore/ParameterSet/src/ANDGroupDescription.h"
 #include "FWCore/Utilities/interface/EDMException.h"
 #include "FWCore/ParameterSet/interface/DocFormatHelper.h"
