+++ FWCore/ParameterSet/src/types.cc
@@ -9,7 +9,7 @@
 #include "FWCore/ParameterSet/interface/types.h"
 
 #include "boost/lexical_cast.hpp"
-#include "FWCore/ParameterSet/interface/split.h"
+#include "FWCore/ParameterSet/src/split.h"
 #include "FWCore/Utilities/interface/Parse.h"
 #include <cctype>
 #include <cstdlib>
